cs-1213 fix bug in search form

+++ eXoApplication/contact/webapp/src/main/java/org/exoplatform/contact/webui/popup/UIAddEditPermission.java
@@ -64,7 +64,7 @@ import org.exoplatform.webui.form.UIFormStringInput;
 
 public class UIAddEditPermission extends UIContainer implements UIPopupComponent {
   public static String[]  BEAN_FIELD = {"viewPermission","editPermission"} ;
-  private static String[] ACTION = {"userId","Edit", "Delete"} ;
+  private static String[] ACTION = {"Edit", "Delete"} ;
   private String groupId_ ;
   private String contactId_ ;
   private boolean isSharedGroup ;
@@ -203,11 +203,15 @@ public class UIAddEditPermission extends UIContainer implements UIPopupComponent
           
           OrganizationService organizationService = 
               (OrganizationService)PortalContainer.getComponent(OrganizationService.class) ;
-            List<User> users = organizationService.getUserHandler().findUsersByGroup(remover).getAll() ;
+          List<User> users = organizationService.getUserHandler().findUsersByGroup(remover).getAll() ;
+          List<String> viewUsers = new ArrayList<String>() ;
+          if (group.getViewPermissionUsers() != null) {
+            viewUsers = Arrays.asList(group.getViewPermissionUsers()) ;
+          }
           for (User user : users)
-            contactService.removeUserShareAddressBook(
+            if (viewUsers.size() > 0 && !viewUsers.contains(user.getUserName() + JCRDataStorage.HYPHEN))
+              contactService.removeUserShareAddressBook(
                 SessionProviderFactory.createSessionProvider(), username, uiForm.groupId_, user.getUserName()) ;
-          
         } else {
           if(group.getViewPermissionUsers() != null) {
             List<String> newPerms = new ArrayList<String>() ;
@@ -235,21 +239,26 @@ public class UIAddEditPermission extends UIContainer implements UIPopupComponent
             SessionProviderFactory.createSessionProvider(), username, uiForm.contactId_) ;
         if (contact.getViewPermissionGroups() != null && Arrays.asList(contact.getViewPermissionGroups()).contains(remover)) {
           removePerGroup(contact, remover) ;
-          
           OrganizationService organizationService = 
               (OrganizationService)PortalContainer.getComponent(OrganizationService.class) ;
-            List<User> users = organizationService.getUserHandler().findUsersByGroup(remover).getAll() ;
+          List<User> users = organizationService.getUserHandler().findUsersByGroup(remover).getAll() ;
+          List<String> viewUsers = new ArrayList<String>() ;
+          if (contact.getViewPermissionUsers() != null) {
+            viewUsers = Arrays.asList(contact.getViewPermissionUsers()) ;
+          }
           for (User user : users)
-            try {
-              contactService.removeUserShareContact(
-                  SessionProviderFactory.createSystemProvider(), username, uiForm.contactId_, user.getUserName()) ;
-            } catch (PathNotFoundException e) {
-              UIApplication uiApp = uiForm.getAncestorOfType(UIApplication.class) ;
-              uiApp.addMessage(new ApplicationMessage("UIAddEditPermission.msg.cannot-deleteShared", null,
-                  ApplicationMessage.WARNING)) ;
-              event.getRequestContext().addUIComponentToUpdateByAjax(uiApp.getUIPopupMessages()) ;
-              return ;
-            }    
+            if (viewUsers.size() > 0 && !viewUsers.contains(user.getUserName() + JCRDataStorage.HYPHEN)) {            	
+              try {
+            	contactService.removeUserShareContact(
+            	  SessionProviderFactory.createSystemProvider(), username, uiForm.contactId_, user.getUserName()) ;
+              } catch (PathNotFoundException e) {
+            	UIApplication uiApp = uiForm.getAncestorOfType(UIApplication.class) ;
+            	uiApp.addMessage(new ApplicationMessage("UIAddEditPermission.msg.cannot-deleteShared", null,
+            	  ApplicationMessage.WARNING)) ;
+            	event.getRequestContext().addUIComponentToUpdateByAjax(uiApp.getUIPopupMessages()) ;
+            	return ;
+              }    
+            }
         } else {
           removePerUser(contact, remover + JCRDataStorage.HYPHEN) ;
           try {
